shorts/requirements.txt
google-genai
python-dotenv
fastapi
uvicorn
gTTS
jinja2
requests
spacy
keybert
sentence-transformers
transformers
scikit-learn
